03/requirements.txt
opencv-python~=4.6.0.66
Pillow~=9.3.0
